services/catalog-collector/requirements.txt
aio-pika==9.3.1
asyncpg==0.29.0
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
beautifulsoup4==4.12.2